refactor: moved makeRequest to own mixin file
