+++ src/mixins/index.js
@@ -11,6 +11,7 @@ const { getDevicePowerUsageRaw } = require('./getDevicePowerUsageRaw');
 const { getDevices } = require('./getDevices');
 const { getFirmwareVersion } = require('./getFirmwareVersion');
 const { getRegion } = require('./getRegion');
+const { makeRequest } = require('./makeRequest')
 const { openWebSocket } = require('./openWebSocket');
 const { saveDevicesCache } = require('./saveDevicesCache');
 const { setDevicePowerState } = require('./setDevicePowerState');
@@ -30,6 +31,7 @@ const mixins = {
   getDevices,
   getFirmwareVersion,
   getRegion,
+  makeRequest,
   openWebSocket,
   saveDevicesCache,
   setDevicePowerState,
